[FEAT] Show System information on Log Screen (#784)
* [Ref/Feat] System Information (#3027)

* Move `getLegendaryVersion`, `getGogdlVersion`, `getNileVersion` and
`getHericVersion` outside of utils.ts

The behavior of everything but `getHeroicVersion` stayed the same.
Heroic's own version codenames will now live inside `package.json` (right next
to the version number) for convenience

* Move `searchForExecutableOnPath` out of utils.ts

This also adds the `hasExecutable` shorthand

* Add a generic "download this file" utility

We've had variations of this in several smaller `util` files, let's go for
making one generic one instead

* Add a generic "run this process" utility

Versions of something like this (`callRunner` for example) have existed as
purpose-built functions, but implementing something new that intents to spawn
processes always needed a new "spawn this process with these arguments and give
me the output" function along with it. Time to change that

* Add a new `systeminfo` utility

This is essentially the `systeminformation` module, but tailor-made for our use
case (no longer getting stuck, gathering more information, respecting
Flatpak paths)

* Add a new "System Information" section to the Settings

This offers the user's system information in a more visual and easy-to-copy way

* Add an Apple logo (for M1/M2/... CPUs)

* Report correct memory use on Linux & macOS

* Add a property describing whether we're in a Flatpak container

* chore: after cherry-pick fixes

* feat: upgrade filesize

* chore: app name

* chore: adjustments and fixes

* [Fix] Several System Information improvements (#3070)

* Report Host OS info in a Flatpak container

On Flatpak, `/run/host/os-release` will exist & contain the host OS'
information. I intended to report this information, but missed a
`break` in the loop (so it'd always find `/etc/os-release`
afterwards and read that out instead)

* Pad PCI IDs with 0

This avoids issues with very old NVIDIA GPUs.
The GPU in question was a Quadro 2000, with an ID of "0dd8". Heroic
searched for just "dd8", which is not in the PCI IDs list

* Make the memory detection more resistant to locale issues

- Certain non-english characters aren't considered "Word characters" in
  Regex, so \S (any non-whitespace) is now accepted
- The `free` command is ran with `LANG=C` to force English output even
  on non-english systems
- If the match fails for any other reason, we simply return 0 bytes
  total/used (which is at least a valid number, even if it doesn't make
  sense)

* Catch `getSystemInfo`-related errors

When initializing Heroic's general / the game's log, we don't *have to*
get system information for anything else to work; we can just report
that there was an error & carry on

* Bump up the max cache time to 30 days

Somewhat of a band-aid fix; ideally we'd be smarter about this caching
in the first place (for example only re-downloading if the user's
hardware changed)

* fix: caching downloadFile

* chore: adjustments and fixes

* Update translation and add macOS version requirement

* chore: remove unused code

* fix: lint

* chore: fix user agent

* Remove handleGoToScreen function from misc.ts

* chore: show hyperplay instead of Software

* fix: lint

---------

Co-authored-by: Mathis Dröge <mathis.droege@ewe.net>
Co-authored-by: Flavio F Lima <flavioislima@users.noreply.github.com>